+++ tests/test_netcdf_monitor.py
@@ -4,6 +4,7 @@ import os
 from datetime import datetime, timedelta
 import numpy as np
 import xarray as xr
+import unittest
 
 random = np.random.RandomState(0)
 
@@ -23,6 +24,67 @@ state = {
         attrs={'units': 'Pa'},
     ),
 }
+
+
+class NetCDFMonitorAliasTests(unittest.TestCase):
+
+    def setUp(self):
+        self.ncfile = 'out.nc'
+
+    def tearDown(self):
+        if os.path.isfile(self.ncfile):
+            os.remove(self.ncfile)
+
+    def store_state_and_check_file(self, aliases):
+        assert not os.path.isfile(self.ncfile)
+        monitor = NetCDFMonitor(self.ncfile, aliases=aliases,
+                                write_on_store=True)
+        monitor.store(state)
+        assert os.path.isfile(self.ncfile)
+
+    def check_nc_var(self, varname, varunit):
+        with xr.open_dataset(self.ncfile) as ds:
+            assert len(ds.data_vars.keys()) == 2
+            assert varname in ds.data_vars.keys()
+            assert ds.data_vars[varname].attrs['units'] == varunit
+            assert tuple(ds.data_vars[varname].shape) == (1, nx, ny, nz)
+
+    def test_keys_string_values_string(self):
+        aliases = {'air_temperature': 'T'}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('T', 'degK')
+        self.check_nc_var('air_pressure', 'Pa')
+
+    def test_keys_nonstring_values_string(self):
+        aliases = {1.0: 'T'}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('air_temperature', 'degK')
+        self.check_nc_var('air_pressure', 'Pa')
+
+    def test_keys_string_values_nonstring(self):
+        aliases = {'air_temperature': 1.0}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('air_temperature', 'degK')
+        self.check_nc_var('air_pressure', 'Pa')
+
+    def test_keys_string_values_emptystring(self):
+        aliases = {'air_temperature': ''}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('air_temperature', 'degK')
+        self.check_nc_var('air_pressure', 'Pa')
+
+    def test_empty_aliases(self):
+        aliases = {}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('air_temperature', 'degK')
+        self.check_nc_var('air_pressure', 'Pa')
+
+    def test_two_aliases(self):
+        aliases = {'air_temperature': 'T',
+                   'air_pressure': 'P'}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('T', 'degK')
+        self.check_nc_var('P', 'Pa')
 
 
 def test_netcdf_monitor_initializes():