+++ HISTORY.rst
@@ -29,6 +29,14 @@ Latest
   Implicit.
 * Added a function reset_constants to reset the constants library to its
   initial state.
+* Shortened the name of tendency diagnostics in TendencyInDiagnosticsWrapper
+* Added "aliases" kwarg to NetCDFMonitor, allowing the monitor to shorten
+  variable names when writing to netCDF
+* Added get_component_aliases() to get a dictionary of quantity aliases from
+  a list of Components (used by NetCDFMonitor to shorten variable
+  names)
+* Added tests for NetCDFMonitor aliases and get_component_aliases()
+
 
 Breaking changes
 ~~~~~~~~~~~~~~~~